move spawn heuristic eval to viewport_blueprint
